refactor: rename llvm_temporary crate to llvm and update build targets
This commit finalizes the integration of the LLVM backend by renaming
the crate and internal modules to more permanent names. It also updates
the build script to include WebAssembly as a supported target.

Changes:
- **Crate Renaming**:
  - Renamed the `llvm_temporary` crate to `llvm` in the workspace
    `Cargo.toml` and the crate's own `Cargo.toml`.
  - Renamed the internal module `llvm_backend` to `backend` for brevity
    and consistency.
- **Integration Updates**:
  - Updated all import paths in `src/cli.rs` and `src/runner.rs` to
    reflect the new crate and module structure.
  - Cleaned up redundant nesting in LLVM-related imports.
- **Build System**:
  - Updated `x.py` to add `wasm32-wasip1` to the list of detected
    targets on Linux.
  - Temporarily disabled the `x86_64-pc-windows-gnu` target in the
    automated build script.

This refactoring marks a transition from a temporary LLVM integration
towards a stable, unified backend structure.
