+++ llvm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "llvm_temporary"
+name = "llvm"
 version = "0.1.0"
 edition = "2021"
 build = "build.rs"
