+++ llvm/src/backend.rs
@@ -14,7 +14,7 @@ use std::path::Path;
 use std::process::Command;
 
 fn ensure_target_dir() {
-    let target_dir = Path::new("target");
+    let target_dir = Path::new("../../target");
     if !target_dir.exists() {
         fs::create_dir_all(target_dir)
             .expect("Unable to create target directory");
@@ -87,7 +87,7 @@ pub fn link_objects(objects: &[String], output: &str, libs: &[String], lib_paths
 pub fn compile_ir_to_img_code(ir: &str, file_stem: &str) -> String {
     ensure_target_dir();
 
-    let target_dir = Path::new("target");
+    let target_dir = Path::new("../../target");
     if !target_dir.exists() {
         fs::create_dir_all(target_dir).expect("Unable to create target directory");
     }
