+++ llvm/src/codegen/consts.rs
