+++ src/runner.rs
@@ -13,8 +13,8 @@ use crate::DebugFlags;
 use ::error::*;
 use ::parser::*;
 use lexer::Lexer;
-use llvm_temporary::llvm_temporary::llvm_backend::*;
-use llvm_temporary::llvm_temporary::llvm_codegen::*;
+use llvm::backend::*;
+use llvm::codegen::*;
 use std::collections::HashSet;
 use std::path::Path;
 use std::{fs, process, process::Command};
